backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
redis[hiredis]==5.2.1
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
fakeredis
typing-extensions>=4.11.0
Faker==33.3.1
google-genai
